refactor: 기존 통합 테스트에 적용
